GH-14: Seperate commands out into own file
